requirements.txt
langchain
langchain-openai
python-dotenv
tiktoken
langchain-text-splitters